backend/requirements.txt
pydantic
python.dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
sarvamai
pypdf
bs4
arxiv
pymupdf
langchain_huggingface
faiss-cpu
langchain-pinecone
langchain_experimental
pdfplumber
flask
flask-cors
requests
werkzeug
pymango
langchain-mongodb
motor
flask[async]
KeyBERT
langchain-milvus
langgraph
tavily-python
chromadb
requests
forex-python
amadeus